+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Create Project `ricaun.Revit.DB.Shape.Tests`
 - Tests for `MaterialUtils`
 - Tests for `Colors`
+- Tests for `TransformUtils`
 
 ## [0.1.0] / 2023-10-01
 ### Quaternion